+++ packages/play/src/main.ts
@@ -9,7 +9,8 @@ import PlusProComponents from '../../plus-pro-components/index'
 import '../../theme-chalk/src/index.scss'
 import App from './App.vue'
 import './style.css'
-import 'element-plus/dist/index.css'
+import 'element-plus/theme-chalk/src/index.scss'
+// import 'element-plus/theme-chalk/dark/css-vars.css'
 
 const app = createApp(App)
 app