+++ packages/play/src/views/form-item.vue
@@ -2,6 +2,7 @@
   <PlusFormItem
     v-for="item in columns"
     :key="item.label"
+    ref="fieldInstance"
     v-model="values[item.prop]"
     v-bind="item"
     @change="(val:any) =>handleChange(val, item.prop)"
@@ -18,7 +19,7 @@ interface RestaurantItem {
 }
 
 const restaurants = ref<RestaurantItem[]>([])
-
+const fieldInstance = ref()
 const createFilter = (queryString: string) => {
   return (restaurant: RestaurantItem) => {
     return restaurant.value.toLowerCase().indexOf(queryString.toLowerCase()) === 0
@@ -38,6 +39,8 @@ const loadAll = () => {
 
 onMounted(() => {
   restaurants.value = loadAll()
+
+  console.log(fieldInstance.value)
 })
 
 const columns: PlusColumn[] = [
@@ -146,6 +149,9 @@ const columns: PlusColumn[] = [
     width: 120,
     prop: 'status',
     valueType: 'select',
+    fieldProps: {
+      multiple: true
+    },
     options: [
       {
         label: '未解决',
@@ -228,6 +234,11 @@ const columns: PlusColumn[] = [
     valueType: 'slider'
   },
   {
+    label: 'text',
+    prop: 'text',
+    valueType: 'text'
+  },
+  {
     label: 'time-picker',
     prop: 'time-picker',
     valueType: 'time-picker'
@@ -255,8 +266,9 @@ const columns: PlusColumn[] = [
 
 const row = {
   name: 'name',
-  status: '1',
+  status: ['0'],
   tag: 'success',
+  text: 'text',
   money: '100',
   progress: 30,
   rate: 4,